More refactor for 311-340 shims.
